# PeaRL_spec/openapi/pearl-api-v1.1.yaml
openapi: 3.1.0
info:
  title: PeaRL API
  version: 1.1.0
  description: API-first risk orchestration platform for autonomous coding and secure/responsible
    delivery.
servers:
- url: http://localhost:8080
  description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^proj_[A-Za-z0-9_-]+$
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 200
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    TaskPacket:
      $ref: ../schemas/context/task-packet.schema.json
...
    JobStatus:
      $ref: ../schemas/workflow/job-status.schema.json
    ReportRequest:
      $ref: ../schemas/workflow/report-request.schema.json
    ReportResponse:
      $ref: ../schemas/workflow/report-response.schema.json
    ErrorResponse:
      $ref: ../schemas/common/error-response.schema.json
  responses:
    BadRequest:
      description: Validation or contract error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/task-packets:
    post:
      tags:
      - TaskPackets
      summary: Generate task-scoped context packet
      operationId: generateTaskPacket
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                schema_version:
                  type: string
                task_type:
                  type: string
                  enum:
                  - feature
                  - fix
                  - remediation
                  - refactor
                  - config
                  - policy
                task_summary:
                  type: string
                affected_components:
                  type: array
                  items:
                    type: string
                change_hints:
                  type: array
                  items:
                    type: string
                environment:
                  type: string
                  enum:
                  - sandbox
                  - dev
                  - pilot
                  - preprod
                  - prod
                trace_id:
                  type: string
                context_budget:
                  type: object
                  additionalProperties: false
                  properties:
                    max_tokens_hint:
                      type: integer
                      minimum: 256
              required:
              - schema_version
              - task_type
              - task_summary
              - environment
              - trace_id
            examples:
              taskPacketReq:
                externalValue: ../examples/task-packets/generate-task-packet.request.json
      responses:
        '201':
          description: Task packet generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPacket'
              examples:
                taskPacket:
                  externalValue: ../examples/task-packets/generate-task-packet.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/projects/{project_id}/reports/generate:
    post:
      tags:
      - Reports
      summary: Generate report
      operationId: generateReport
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
            examples:
              reportReq:
                externalValue: ../examples/reports/generate-report.request.json
      responses:
        '200':
          description: Report generated or returned if synchronous
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              examples:
                reportResp:
                  externalValue: ../examples/reports/generate-report.response.json
        '202':
          description: Report job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'